+++ tools/deploy.ps1
@@ -22,6 +22,7 @@ $items = @{
                   @{ Src = "$root\addons\DLSS5NR01\build\Release\fsr\amd_fidelityfx_dx12.dll"; Rel = 'fsr\amd_fidelityfx_dx12.dll' },
                   @{ Src = "$root\addons\DLSS5NR01\external\fsr4\amd_fidelityfx_dx12.dll"; Rel = 'runtimes\FSR\0dd77d9c\amd_fidelityfx_dx12.dll' },
                   @{ Src = "$root\addons\DLSS5NR01\products\FSR3UPSC\FSR4-ABOUT.txt"; Rel = 'runtimes\FSR\0dd77d9c\ABOUT.txt' },
+                  @{ Src = "$root\addons\DLSS5NR01\third_party\ffx4\LICENSE.txt"; Rel = 'runtimes\FSR\0dd77d9c\LICENSE.txt' },
                   @{ Src = "$root\addons\DLSS5NR01\third_party\ffx\LICENSE.txt"; Rel = 'AMD-FidelityFX-LICENSE.txt' },
                   @{ Src = "$root\addons\DLSS5NR01\LICENSE"; Rel = 'LICENSE.txt' }) }
 }